consumer/requirements.txt
requests
beautifulsoup4
kafka-python
pandas
pyarrow
azure-storage-blob
python-dotenv
selenium
